current_requirements.txt
pip==24.0
setuptools==65.5.0
wheel==0.38.4
cmake==4.2.1
packaging==25.0

six==1.17.0
typing_extensions==4.13.2
importlib_metadata==8.5.0
importlib_resources==6.4.5
zipp==3.20.2

certifi==2024.8.30
charset-normalizer==3.4.4
idna==3.11
urllib3==2.2.3
requests==2.32.4
oauthlib==3.3.1
requests-oauthlib==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
google-auth==2.47.0
google-auth-oauthlib==1.0.0

filelock==3.16.1
fsspec==2025.3.0
networkx==3.1
numpy==1.24.4
mpmath==1.3.0
sympy==1.13.3

pillow==10.4.0
imageio==2.35.1
pygame==2.6.1
pyglet==1.5.21

matplotlib==3.7.5
contourpy==1.1.1
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.7
pyparsing==3.1.4

python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3

torch==2.2.2
torchvision==0.17.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
grpcio==1.70.0
protobuf==3.20.3
absl-py==2.3.1
Markdown==3.7
MarkupSafe==2.1.5
Werkzeug==3.0.6
Jinja2==3.1.6
cloudpickle==3.1.2

gym==0.21.0
gymnasium==1.1.1
nes-py==8.2.1
gym-super-mario-bros==7.3.0

Shimmy==2.0.0
stable-baselines3==1.6.0
procgen==0.10.7
Farama-Notifications==0.0.4

colorama==0.4.6
icecream==2.1.9
executing==2.2.1
asttokens==3.0.1
Pygments==2.19.2
tqdm==4.67.1
pandas==2.0.3
